requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.89.1
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httptools==0.5.0
idna==3.4
psycopg2-binary==2.9.5
pydantic==1.10.4
pyenchant==3.2.2
python-dotenv==0.21.1
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==2.0.1
starlette==0.22.0
typing_extensions==4.4.0
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
